+++ package.json
@@ -27,7 +27,7 @@
   "scripts": {
     "start": "run-p client storybook",
     "client": "react-scripts start",
-    "server": "nodemon server/sslserver.js",
+    "server": "nodemon server/server.js",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
