--- server/server.cert
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDsjCCApoCCQDcOaCQsdCi7TANBgkqhkiG9w0BAQsFADCBmjELMAkGA1UEBhMC
-ZGUxGzAZBgNVBAgMEnNjaGxlc3dpZy1ob2xzdGVpbjENMAsGA1UEBwwEa2llbDEg
-MB4GA1UECgwXdmFjY2luYXRpb24tY2VydGlmaWNhdGUxDzANBgNVBAsMBnNlcnZl
-cjEOMAwGA1UEAwwFbG9jYWwxHDAaBgkqhkiG9w0BCQEWDXRlc3RAdGVzdC5jb20w
-HhcNMTkxMjEyMjE1MzQwWhcNMjAwMTExMjE1MzQwWjCBmjELMAkGA1UEBhMCZGUx
-GzAZBgNVBAgMEnNjaGxlc3dpZy1ob2xzdGVpbjENMAsGA1UEBwwEa2llbDEgMB4G
-A1UECgwXdmFjY2luYXRpb24tY2VydGlmaWNhdGUxDzANBgNVBAsMBnNlcnZlcjEO
-MAwGA1UEAwwFbG9jYWwxHDAaBgkqhkiG9w0BCQEWDXRlc3RAdGVzdC5jb20wggEi
-MA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCuJ3yc5L26j2Me29WDDWIDi+tY
-XQCuT771s/9FeCB/6sGo/9TRdJzwlz+HvdMQQwgWH/k7HqFle/0a1j1lzdgrbf8E
-mRUgdBjSw9GduE9BJGROsoXNnEHUYVyao47btsrO6EtUUlz2Z44ucfRENTq75C9K
-ZlSOAyDVcP3D7mvrwv9q/tQSWh0gdAqHAf0e3PImxfWrzexb7d32J0gusXCDaokT
-anoG1+ONxW8rkSITQ3z7Qz+tI6wiCd0uVsGUlYeT4iRSD7+AXy+N8gcyPAS6gBze
-J3juCIR7Ry3iAqO9b37wr1gdl6R2gE7TIGboiqS3FbEPeBzLVt77UYLHw2flAgMB
-AAEwDQYJKoZIhvcNAQELBQADggEBACIAgbWSf0kovjTo2s7J3NharKsVSXYCznYQ
-f5twYqp4SWsZs3G7lYGjKpr/n803B/UrXYQvBGHrvzTg/q4fDfmI+7P9TNlSbeL+
-gHX5Dst3T09fBI1DWFqRKE98Im0zo5J2pmx8VdJmhmWMh3SMMQhCsvB+q+9dTfTE
-/XhCdfjTp+OnU32OZ8dMXntxGQYUxkatrkuzytjaQTUo0Kc4CZB1gBP7HMdL8c1D
-7DryDINuu++MqsYOhZDo1JP9R5lkcdmMxy0/vgmfAvpU3rjVAq8tJIW22qaBNYfh
-bH/8LTbGgP161/E93uDKXoDvUulvQKE9u9xPqnKpAWIo9lRp1sE=
------END CERTIFICATE-----
