+++ server/server.js
@@ -1,6 +1,6 @@
-const updateVaccinations = require('./updateVaccinationsOpen')
+const updateVaccinations = require('./controllers/updateVaccinationsOpen')
 const userFormatter = require('./controllers/Formatter/userFormatter')
-const createVaccinationsMadeFromVaccine = require('./createVaccinationsMadeFromVaccine')
+const createVaccinationsMadeFromVaccine = require('./controllers/createVaccinationsMadeFromVaccine')
 const toDateObject = require('./controllers/Time/toDateObject')
 
 const express = require('express')
@@ -11,7 +11,7 @@ const cors = require('cors')
 const port = 3338
 
 const admin = require('firebase-admin')
-let serviceAccount = require('./medical-assistant-19fc3-cee7dfd4e3aa.json')
+let serviceAccount = require('./.env/medical-assistant-19fc3-cee7dfd4e3aa.json')
 
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount),
@@ -19,26 +19,24 @@ admin.initializeApp({
 
 let db = admin.firestore()
 
-//check for CORS func
 app.use(cors())
 app.use(express.json())
 app.set('json spaces', 2)
 
 app.use(function(req, res, next) {
-  res.header('Access-Control-Allow-Origin', 'http://localhost:3000') // update to match the domain you will make the request from
+  res.header('Access-Control-Allow-Origin', 'http://localhost:3000')
   res.header(
     'Access-Control-Allow-Headers',
     'Origin, X-Requested-With, Content-Type, Accept, authorization'
   )
-  //res.header(res.header('Access-Control-Allow-Methods', 'PATCH'))
   next()
 })
 
 https
   .createServer(
     {
-      key: fs.readFileSync('./server.key'),
-      cert: fs.readFileSync('./server.cert'),
+      key: fs.readFileSync('./.env/server.key'),
+      cert: fs.readFileSync('./.env/server.cert'),
     },
     app
   )
@@ -51,6 +49,7 @@ https
 app.get('/api/:user', function(req, res) {
   const { user } = req.params
   const idToken = req.headers.authorization
+  console.log('request', req)
   admin
     .auth()
     .verifyIdToken(idToken)
@@ -118,8 +117,6 @@ app.patch('/api/settings/:user', (req, res) => {
 })
 
 app.patch('/api/:user', (req, res) => {
-  console.log('PATCH')
-  console.log(req.body)
   const { user } = req.params
   const idToken = req.headers.authorization
   admin
@@ -127,9 +124,7 @@ app.patch('/api/:user', (req, res) => {
     .verifyIdToken(idToken)
     .then(function(decodedToken) {
       let uid = decodedToken.uid
-      console.log('uid', uid)
       if (uid === user) {
-        console.log(req.body)
         db.collection('vaccines')
           .doc(req.body.sticker)
           .get()
@@ -144,9 +139,6 @@ app.patch('/api/:user', (req, res) => {
                 vaccine,
                 req
               )
-              console.log(newVaccinations)
-              console.log('---------------------')
-              console.log('new vaccinations', newVaccinations)
               db.collection('users')
                 .doc(user)
                 .update({
