--- server/server.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCuJ3yc5L26j2Me
-29WDDWIDi+tYXQCuT771s/9FeCB/6sGo/9TRdJzwlz+HvdMQQwgWH/k7HqFle/0a
-1j1lzdgrbf8EmRUgdBjSw9GduE9BJGROsoXNnEHUYVyao47btsrO6EtUUlz2Z44u
-cfRENTq75C9KZlSOAyDVcP3D7mvrwv9q/tQSWh0gdAqHAf0e3PImxfWrzexb7d32
-J0gusXCDaokTanoG1+ONxW8rkSITQ3z7Qz+tI6wiCd0uVsGUlYeT4iRSD7+AXy+N
-8gcyPAS6gBzeJ3juCIR7Ry3iAqO9b37wr1gdl6R2gE7TIGboiqS3FbEPeBzLVt77
-UYLHw2flAgMBAAECggEAKjWP0kbuZSUTlQsa+KRpuQpUhBombBkP42U43ZF77wTq
-4xb+cbWdxAIkd7XeTfaxIwuk0FtgKsSeV77NKrt1RTFnuamQoUkCKzMivuNI8jTr
-ytch7u+MVPk0BJbXrRDknrlWaJ19worevwsHdPChx6BKhfG/VJwVfSXhm3dR2NYH
-cVQxlrD/WCezgtvTf6vEB3Qi3bJLpI6I81WwwmPymmCt84B7hJ9ZIET1CCk2g8nQ
-tQm5tYjPmfnUAHE45YOe1pPmmpy+pfhzQfIH7hwbtLrk5FZPkvqPwU3Y8xE/G2PV
-xXRW0GK3GEUVTxqWtxkelEVNB803p0zME+A0VB2WkQKBgQDl2JEcrlWXMupP4DY/
-yzKoe87YYRMJLB2juuiCHlR6qNRsi7gokTw2vGPDxEPWGjvDeeM5v8Yazm9RCIkT
-+vPu24bjPqXFVh7gOOXjBaG3jv68P9y64CFI1FqlOGMqrEtUUmrJgMGNjHH7Jn2t
-zg6AlOjN17smoBr97oHb/z4vawKBgQDB+J2BEAoZWKI7aoePPaES4hUN60yw67Ni
-4VXJerkReQ8MBgNvGLqXWz5iiY/6NkTU7CiJc95Kbpet09h8sJLDbTUzC8C64Pdg
-oqU4m63vSrrrr9cGXa3xIUHaMvexmDeubuorYrn9d00Xt3fCKoo7g0Phpxn/L5xE
-k6wgbZcp7wKBgQDI2V+tNXabwqVYmXAufO1Op6oQEXwl9sQL0IffNIvUB/ZgyBNs
-d8WreyKe7sHstU3mOnvief+uaxiBMjk/GXvSDUmpCJSpheFDRNU2djRkB/P4IsJL
-FxJe5xoy6LxXjOnizIZ+gSIiBEmSE4edusOXfwiU3+CuFxifL62QxlHb8wKBgAQi
-fZEDu4NgzUyA79TQ1n8ljJdYXbEjxRYlgYOi9xD1SKnT8vzKRSf2zPSixAvsXjk/
-s3mCm8lwZBFTg7jctrQtvsWQyZHyO7O2o9j6EJIaQLPHcsjy9ENAlzvEfF0vxngU
-ay9zkk1h3qXTxlzANWQhXzAPNNGovScFEtU6IM/HAoGBAKjDalxAKtrqT3vovK0I
-bwiwEGFcXTfllzu1xX5VnTfQMfkmMfiu40I5bcr67SY3QWilMq3HKwZpV7L8TgC7
-JNCRuAFg1/Ox/oT3Srrs1fBsY1RERKPyHYj2rrcoqt5mMZfcfqjYXZqdrFmJAe+5
-+Ei+NLlGWCezR9P1dseuvb4L
------END PRIVATE KEY-----
