+++ src/App.js
@@ -12,6 +12,7 @@ export default function App() {
     auth.onAuthStateChanged(user => {
       if (user) {
         setUser(user)
+        console.log(user)
       } else {
         setUser(user)
       }
